move modules into /modules
