+++ src/modules/auth/dto/auth-credentials.dto.ts
@@ -1,10 +1,12 @@
 import { IsString, Matches, MaxLength, MinLength } from 'class-validator';
 
+// don't know why call it DTO
 export class AuthCredentialsDto {
   @IsString()
   @MinLength(4)
   @MaxLength(20)
   username: string;
+
   @IsString()
   @MinLength(8)
   @MaxLength(20)
