+++ src/modules/comment/dto/create-comment.dto.ts
@@ -3,6 +3,7 @@ import { IsNotEmpty, IsNumber, IsOptional } from 'class-validator';
 export class CreateCommentDto {
   @IsNotEmpty()
   description: string;
+
   @IsOptional()
   @IsNumber()
   postId: number;
